pyproject.toml
[project]
name = "openadapt-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.7",
    "peft>=0.18.0",
    "pillow>=12.0.0",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
]
